chore: handle balance to invest flow
