+++ src/commands/invest/token/processor.ts
@@ -6,14 +6,10 @@ import { formatDigit } from "utils/defi"
 import { VERTICAL_BAR } from "utils/constants"
 import { ApiEarningOption, ApiPlatform } from "types/krystal-api"
 import { flatten } from "lodash"
-import { ButtonInteraction, CommandInteraction } from "discord.js"
 
 type EarningPlatform = ApiPlatform & { chainName: string }
 
-export async function renderInvestToken(
-  i: CommandInteraction | ButtonInteraction,
-  token: string
-) {
+export async function renderInvestToken(token: string) {
   let tokenData = [] as EarningPlatform[]
   const { result, ok } = await CacheManager.get({
     pool: "invest",
